requirements.txt
python-chess==0.24.2
